requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Telegram Bot API (aiogram v3 with Redis FSM storage)
aiogram==3.13.0
telethon>=1.36.0

# Task queue
celery==5.4.0
redis==5.1.1
flower==2.0.1

# YouTube Data API v3
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Video processing (requires system FFmpeg installed)
ffmpeg-python==0.2.0

# Database
supabase==2.9.1

# Data / Export
pandas==2.2.3
openpyxl==3.1.5

# HTTP client
httpx==0.27.2

# Config
python-dotenv==1.0.1
pydantic-settings==2.5.2

# Fuzzy matching
rapidfuzz==3.10.0

# Utilities
python-multipart==0.0.12
aiofiles==23.2.1
Pillow==11.1.0

# Tests
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
